rename some python files
