+++ app/core/clauses/clause.py
