+++ app/core/clauses/match_clause.py
@@ -1,6 +1,6 @@
 from typing import Dict, List, Tuple
 from dataclasses import dataclass
-from app.core.clauses.Clause import Clause
+from app.core.clauses.clause import Clause
 
 @dataclass
 class NodePattern:
