+++ app/core/schema/edge.py
