+++ app/core/translator/query_translator.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import List
 
-from app.core.clauses.Clause import Clause
+from app.core.clauses.clause import Clause
 
 class QueryTranslator(ABC):
