+++ app/core/translator/question_translator.py
